型定義を別ファイルへ移動
