+++ src/features/core/extract/types.ts
@@ -1,0 +1,27 @@
+import type {
+  MapBatchReadResult,
+  ReadSystemResult,
+  DataReadErrorItem,
+} from "@RpgTypes/fileio";
+import type {
+  ExtractedDataBundle,
+  ExtractedCommonEventText,
+  ExtractedBattleEventText,
+  ExtractedMapTexts,
+} from "./text";
+import type { ExtractedSystemTexts } from "./text/system";
+
+export interface ExtractedRawGameDataTextValue {
+  mainData: ExtractedDataBundle;
+  eventData: {
+    commonEvents: ExtractedCommonEventText[];
+    troops: ExtractedBattleEventText[];
+  };
+  mapFiles: MapBatchReadResult<ExtractedMapTexts>;
+  system: ReadSystemResult<ExtractedSystemTexts>;
+}
+
+export interface ExtractedRawGameDataTexts {
+  value: ExtractedRawGameDataTextValue;
+  errors: DataReadErrorItem[];
+}
